requirements.txt
torch
numpy
opencv-python
imutils
matplotlib
torchvision
pillow
streamlit